+++ R/hbam.R
@@ -84,6 +84,10 @@
 #' # Obtaining posterior summaries for the latent stimulus positions:
 #' theta_est <- get_est(fit_hbam_mini, par = "theta")
 #'
+#' # Obtaining posterior summaries for the latent respondent positions
+#'   # in a format that matches the rows in the original dataset:
+#' chi_est <- get_est(fit_hbam_mini, par = "chi", data = dat, format_orig = TRUE)
+#'
 #' # Fitting the FBAM_MULTI_NF model with self-placements as group_id:
 #' fit_fbam_multi_nf <- hbam(self, stimuli, group_id = self, model = "FBAM_MULTI_NF",
 #'                        warmup = 500, iter = 1000, chains = 2)